+++ include/cryptoplus/x509/certificate_request.hpp
@@ -322,7 +322,7 @@ namespace cryptoplus
 		{
 			const unsigned char* pbuf = static_cast<const unsigned char*>(buf);
 
-			return take_ownership(d2i_X509_REQ(NULL, &pbuf, buf_len));
+			return take_ownership(d2i_X509_REQ(NULL, &pbuf, static_cast<long>(buf_len)));
 		}
 		inline certificate_request::certificate_request()
 		{
